requirements.txt
discord.py
fastapi
uvicorn
python-dotenv
python-multipart
httpx
pyttsx3
aiohttp